fixed add function
