+++ Assignment2_TDD_Fleet/Assignment2_TDD_Fleet/Assignment2_TDD_Fleet.csproj
@@ -61,6 +61,7 @@
     <Compile Include="AddVehicle.xaml.cs">
       <DependentUpon>AddVehicle.xaml</DependentUpon>
     </Compile>
+    <Compile Include="CarManagement.cs" />
     <Compile Include="FuelPurchase.cs" />
     <Compile Include="Journey.cs" />
     <Compile Include="Service.cs" />
