+++ Assignment2_TDD_Fleet/Assignment2_TDD_Fleet/CarManagement.cs
@@ -1,0 +1,53 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Assignment2_TDD_Fleet
+{
+    public class CarManagement
+    {
+        private FuelPurchase fuelPurchase;
+
+        /**
+         * Class constructor specifying name of make (manufacturer), model and year
+         * of make.
+         * @param manufacturer
+         * @param model
+         * @param makeYear
+         */
+       
+
+        /*
+        public Vehicle(string manufacture, string model, string makeYear)
+        {
+            this.CarManufacture = manufacture;
+            this.CarModel = model;
+            this.CarYear = makeYear;
+            fuelPurchase = new FuelPurchase();
+        }
+        */
+        /**
+         * Prints details for {@link Vehicle}
+         */
+         /*
+        public void printDetails()
+        {
+            Console.WriteLine("Vehicle: " + CarYear + " " + CarManufacture + " " + CarModel);
+            // TODO Display additional information about this vehicle
+        }
+        */
+
+        // TODO Create an addKilometers method which takes a parameter for distance travelled 
+        // and adds it to the odometer reading. 
+
+        // adds fuel to the car
+        /*
+        public void addFuel(double litres, double price)
+        {
+            fuelPurchase.purchaseFuel(litres, price);
+        }
+        */
+    }
+}
