+++ Assignment2_TDD_Fleet/Assignment2_TDD_Fleet/Vehicle.json
@@ -1,0 +1,1 @@
+[{"vehicleListChanged":false,"Id":1,"RegistrationID":"1828920","CarManufacture":"BMW","CarModel":"Z5","CarYear":"2019","FuelType":"Petrol","TankCapacity":"7.5","VehicleOdometer":"10020"},{"vehicleListChanged":false,"Id":2,"RegistrationID":"2829121","CarManufacture":"Toyota","CarModel":"Innova","CarYear":"2014","FuelType":"Petrol","TankCapacity":"20","VehicleOdometer":"20001"},{"vehicleListChanged":false,"Id":0,"RegistrationID":"test","CarManufacture":"","CarModel":"","CarYear":"","FuelType":"","TankCapacity":"","VehicleOdometer":""}]
